+++ src/components/global/utils/eventApiUtils.js
@@ -3,12 +3,11 @@ import {
   BASE_URL,
   ACCESS_TOKEN,
   REFRESH_TOKEN,
-  TOKEN_EXPIRES_IN,
   LAST_EVENT_ID,
 } from '../constants/index.ts';
 import { showSuccessToast } from '../../ui/toast/toast.tsx';
-import { setTokens } from '../../../store/actions/authActions.js';
-import { store } from '../../../store/index.js';
+import { reissueTokens } from './reissueTokenUtils.js';
+
 let isConnected = false;
 
 export const initializeSSE = (queryClient) => {
@@ -16,7 +15,7 @@ export const initializeSSE = (queryClient) => {
   isConnected = true;
 
   const accessToken = localStorage.getItem(ACCESS_TOKEN);
-  let lastEventId = null;
+  let lastEventId = localStorage.getItem(LAST_EVENT_ID);
 
   if (!accessToken) {
     console.log('토큰이 없어서 연결을 시작할 수 없습니다.');
@@ -38,7 +37,7 @@ export const initializeSSE = (queryClient) => {
     },
     heartbeatTimeout: 3660000,
   });
-  console.log(sse);
+
   sse.addEventListener('connect', (event) => {
     queryClient.invalidateQueries(['notifications']);
 
@@ -80,35 +79,8 @@ export const initializeSSE = (queryClient) => {
     console.log('onerror', event);
     sse.close();
     isConnected = false;
-    reissueTokens();
+    reissueTokens(queryClient);
   };
 
   return sse;
-};
-
-const reissueTokens = async () => {
-  try {
-    const response = await fetch(`${BASE_URL}/api/auth/reissue`, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem(ACCESS_TOKEN)}`,
-        'Authorization-refresh': `Bearer ${localStorage.getItem(REFRESH_TOKEN)}`,
-      },
-    });
-
-    if (response.ok) {
-      const newAccessToken = response.headers.get('Authorization');
-      const newRefreshToken = response.headers.get('Authorization-refresh');
-      localStorage.setItem(ACCESS_TOKEN, newAccessToken);
-      localStorage.setItem(REFRESH_TOKEN, newRefreshToken);
-      const expiresAt = new Date().getTime() + TOKEN_EXPIRES_IN * 1000;
-      store.dispatch(setTokens(newAccessToken, newRefreshToken, expiresAt));
-      setTimeout(initializeSSE, 1000);
-    } else {
-      throw new Error('토큰 재발급 실패');
-    }
-  } catch (error) {
-    console.error('토큰 재발급 오류:', error);
-    setTimeout(reissueTokens, 5000);
-  }
 };
